db · Frente C · troceo 1/N · extraer helpers de blob/documento → services/db/documents.ts
Mover-no-reescribir: getDocumentBlob, downloadBlob, saveDocumentWithBlob,
deleteDocumentAndBlob salen de db.ts a services/db/documents.ts (copia literal).
db.ts los importa (downloadBlob se usa internamente en el export de snapshot) y los
re-exporta → cero cambios para los consumidores. Import diferido (initDB se invoca
dentro de funciones async) → sin ciclo en runtime; verificado porque el test
network-first + initDBBloque3 + backfillDocumentoFirmado cargan db.ts a través del
grafo circular y pasan.

db.ts 4553→4466 líneas. tsc 0.
