+++ src/services/db/documents.ts
@@ -1,0 +1,102 @@
+// Frente C · troceo de db.ts (mover-no-reescribir).
+// Utilidades de blob/documento extraídas literalmente de db.ts (H0.4).
+// Se importan `initDB`/`Document` de `../db`; la referencia es diferida
+// (initDB se invoca dentro de las funciones async, nunca en load) → sin ciclo
+// en runtime. `db.ts` re-exporta estas funciones para no tocar a los consumidores.
+
+import { initDB } from '../db';
+import type { Document } from '../db';
+
+// Blob storage and download utilities (H0.4 requirement)
+export const getDocumentBlob = async (id: number): Promise<Blob | null> => {
+  try {
+    const db = await initDB();
+    const doc = await db.get('documents', id);
+    return doc?.content || null;
+  } catch (error) {
+    console.error('Error retrieving document blob:', error);
+    return null;
+  }
+};
+
+export const downloadBlob = (blob: Blob, filename: string): void => {
+  try {
+    // For iOS/Safari compatibility, try dataURL method first for smaller files
+    const isIOSSafari = /iPad|iPhone|iPod/.test(navigator.userAgent) && !(window as any).MSStream;
+
+    if (isIOSSafari && blob.size < 50 * 1024 * 1024) { // < 50MB for iOS Safari
+      const reader = new FileReader();
+      reader.onload = () => {
+        const dataUrl = reader.result as string;
+        const a = document.createElement('a');
+        a.href = dataUrl;
+        a.download = filename;
+        document.body.appendChild(a);
+        a.click();
+        document.body.removeChild(a);
+      };
+      reader.readAsDataURL(blob);
+    } else {
+      // Standard blob URL method for other browsers
+      const url = URL.createObjectURL(blob);
+      const a = document.createElement('a');
+      a.href = url;
+      a.download = filename;
+      document.body.appendChild(a);
+      a.click();
+      document.body.removeChild(a);
+      URL.revokeObjectURL(url);
+    }
+  } catch (error) {
+    console.error('Error downloading file:', error);
+    throw new Error('No se pudo descargar el archivo');
+  }
+};
+
+export const saveDocumentWithBlob = async (document: Omit<Document, 'id'> & { id?: number }): Promise<number> => {
+  try {
+    const db = await initDB();
+
+    // Ensure proper type detection for ZIP files
+    if (!document.type || document.type === '') {
+      const filename = document.filename.toLowerCase();
+      if (filename.endsWith('.zip')) {
+        document.type = 'application/zip';
+      } else {
+        document.type = 'application/octet-stream';
+      }
+    }
+
+    // Add metadata for blob storage
+    const docWithMetadata = {
+      ...document,
+      metadata: {
+        ...document.metadata,
+        createdAt: new Date().toISOString(),
+        blobStored: true,
+      }
+    };
+
+    if (document.id) {
+      await db.put('documents', docWithMetadata as Document);
+      return document.id;
+    } else {
+      const id = await db.add('documents', docWithMetadata);
+      return id as number;
+    }
+  } catch (error) {
+    console.error('Error saving document with blob:', error);
+    throw new Error('No se pudo guardar el documento');
+  }
+};
+
+export const deleteDocumentAndBlob = async (id: number): Promise<void> => {
+  try {
+    const db = await initDB();
+    await db.delete('documents', id);
+    // The blob is automatically deleted with the document record
+  } catch (error) {
+    console.error('Error deleting document and blob:', error);
+    throw new Error('No se pudo eliminar el documento');
+  }
+};
